Use `auto` for `QTE_D` and friends
Use C++11 `auto` keyword to implement `QTE_D` and related macros. This
removes the need to provide the class name, and to have separate flavors
for use in `const` methods, making them much more convenient to use. For
backwards compatibility, the macros now take variable arguments (which
are ignored) so that existing users do not need to be changed.