Add compiler configuration logic
Add a bunch of logic to set up the compiler configuration the way we
want/need. In particular, this turns on C++11 support and checks for
some required language features, and also turns on a whole bunch of
warnings that we would like enabled by default, some of which are
promoted to errors.