Fix documentation paths
Fix how we specify the path to resources used to generate the
documentation. Remove the unneeded `sdk` from the documentation output
path. (This is a vestige left over from ViViA, which also has user
manuals for the applications, in addition to SDK documentation. Since
qtExtensions does not provide end user applications, there is no need
for this separation.)