Create plugin for Designer
Create a Designer plugin that allows using most of the qtExtensions
widgets directly. Thanks to a base interface class and some clever
template magic, the interfaces can be header-only (some function bodies
are still required, but are small enough that making them inline is not
unreasonable, especially as these are leaf headers only used to build
the plugin itself and are not public).

Also, fix a small oversight in qtSqueezedLabel so that its properties
are properly understood by the metatype system (i.e. so that they show
up and can be edited in Designer).

Note also that this is starting to use the new code style; four space
indent, Allman braces, no `this->`, and members like `m_foo`. The goal
is to eventually convert the entire code base to this style, but there
is no immediate plan for doing so, besides new code (i.e. new source
files) using the new style.