Propagate language-level flag to users
Modify how we turn on flags so that we can identify the language-level
flag. Set this as an interface compile option on our library, so that
users will get the flag automatically.

This is a little heavy handed, but will allow us to use modern C++
features in our headers (e.g. explicitly defaulted/deleted members)
without requiring users to reinvent logic to ensure an appropriate
language-level is specified.