Fix "public" conversion warnings
Ad explicit type conversions in some STL helpers. Add suffixes so that
we aren't egregiously mixing floats and doubles. This fixes
`-Wconversion` warnings in headers, so that users that use this warning
flag (and aren't treating our headers as system headers) aren't spammed
with warnings that really shouldn't exist.

There are a bunch of `-Wconversion` warnings in our sources as well that
really ought to also be fixed, but that can happen some other time.