Omit space between template `>`s
Add check that compiler supports omitting space between `>`s in nested
template argument lists (as introduced/required by C++11). Do so in the
small number of instances where we have nested templates.

This is not expected to change our compiler support materially (we
already require partial C++11 support, and GCC 4.4 (Red Hat 6) supports
this as well as the other features we require), so the check is expected
to be more a matter of paranoia. This is a convenient feature to use,
however, as it helps with code style consistency.