Make `qtUiState`s helper item class public
Move the `Item` helper class from `qtUiStatePrivate` to `qtUiState`, and
into a public header. Although this is something of a leak of
implementation details that could possibly come back to bite us later
(if we need to change those details), it's extremely useful for users
that need to persist something that qtUiState doesn't directly provide
to be able to make use of this helper class.