Add support for entity references to qtSax
Add `qtSaxEntity` and `qtSaxText` classes. The former is a simple
wrapper for writing entity references, much like the element and
attribute wrappers. The latter can be used to simply wrap text, but its
main purpose is to wrap (as a convenience) text that contains entities
that should be emitted as actual entities rather than being escaped.