Add check for C++11 initialization features
Add a feature test for uniform initialization and type elision in
returns and function parameters. As with out other C++11 features, this
is supported by GCC 4.4. It's also terribly convenient for passing
temporaries of mostly-aggregate types like QRect (especially if they
have longer names), or for eliding the type name when returning a
default-constructed object.