`for each` -> `foreach`
Rename new iteration helper. Apparently, MSVC performs syntax checks on
code BEFORE THE PREPROCESSOR RUNS, which causes it to reject `for each`
due to missing `in`, even though the preprocessed code would replace the
`each`. The new version is less optimal, but should work on Microsoft's
broken compiler.

Note that this hides `foreach` from Qt. The Qt version, which makes a
copy of the container to be iterated over, is made available via
`foreach_copy`. This is somewhat ugly, as it is a potential (silently)
breaking change, although most users do not need the additional safety
and will see a silent (small) improvement as a result. Unfortunately,
there are not many "pretty" options available for a new pseudo-keyword.