Use uniform initialization in `qtEnumerate`
Modify `qtEnumerate` to use uniform initialization. This was originally
being held back, but we actually already don't support VS2010 (due to
explicitly defaulted/deleted members), and have decided that we are
going to officially drop that as a supported compiler. The remaining
supported compilers all do support this feature, and using it makes the
code prettier. (A feature test is pending on a separate branch that also
updates a wider assortment of the code base to use this feature.)