Use RPATH on OS/X
Turn on RPATH's on OS/X unconditionally. This should give more UNIX-like
behavior as far as libraries that Just Work, path-wise, from a build
directory. More to the point, it silences a warning when using newer
versions of CMake which default to this behavior. (We could also bump
the minimum required CMake version, but we like backward compatibility,
and for now, this doesn't seem like sufficient reason to start requiring
3.x.)