Add `qtIndexRange` helper
Add a helper to enumerate the indices in an index range (i.e. in the
half-open range [, count)). This is immensely useful for iterating over
random access lists, especially as it allows type deduction (via `auto`)
of the index type (which fixes a frequent error made writing such
loops, and is much less ugly than using `decltype`).