Set environment when running `qte-amc`
Modify `qte_amc_wrap_ui` to set the environment (`PATH`, specifically)
when running `qte-amc` on Windows, when sufficiently new CMake is used.
This should make it easier to build, since running `qte-amc`, which is
normally done at build time by projects using it, requires that the Qt
libraries are in the library path.