Wrap qte-amc for a custom environment on OS X
OS X has a similar problem to Windows in that Qt or its dependencies may
not be found when running qte-amc.  Generally this can be fixed by
setting DYLD_FALLBACK_LIBRARY_PATH.  The problem arises on OS X 10.11
(El Capitan) and above in which the system integrity protection feature
clears the DYLD_* environment variables.  This patch uses a similar
trick already applied for Windows to set the environment before calling
qte-amc.