Add `qtGet` helper
Add a helper function to get an item from an associative container.
Specifically, this returns a pointer to the item, if it exists, and
nullptr otherwise. This makes it easy to avoid a double-lookup of the
item, or a call to `find()` immediately followed by testing the validity
of the iterator.