Use only one label
Right now, we are not using both the name and description labels. Since
there has been some debate over the correct formatting, bypass the issue
by instead only having one label. Most of the use cases of having
multiple labels (in whatever layout) can be accomplished with rich text
instead, and if we really need to, we might add the ability to use a
custom widget for the task label, or expand the API again.

This avoid having the API be overly specified for the early
implementation, since we can always make it more complex later; taking
away complexity is much more awkward.