Fix Designer plugin for Qt5
Update how the Designer plugin is built to work with the new Qt5 system
of creating plugins, and to work around moc not being able to find the
include directories when invoked manually. (I suppose the latter is not
really a bug, since the "fix" is to tell it what target will eventually
incorporate the generated files, so that it can find the necessary
include directories. Presumably this is not needed for Qt4 because Qt's
include directories are being added globally, which is the wrong way to
do things.)