Implement natural sorting
Add an implementation of "natural" sorting. The objective is to be able
to "naturally" sort strings which contain numbers, e.g. to sort "x5"
before "x22". This handles non-ASCII digits as well as numbers of any
length. However, it has inevitable limitations, which are described in
the documentation.

For now, sorting of non-numeric potions is strictly lexicographical.
Case insensitivity or locale-aware sorting may be added at some point in
the future. (The comparator is set up as a functor rather than a
function in order to facilitate this possibility in a hopefully source
compatible manner.)