Implement `with` and `using` pseudo-keywords
Add new pseudo-keywords `with` and `using`, which create a scope within
which a declaration is active. Reimplement `synchronized` in terms of
these. The former takes an expression, which is assigned to an
"anonymous", unused variable, while the latter takes a full declaration.
(Unfortunately there is no way I can see to write a preprocessor macro
that would permit either form and still DTRT.)

These will be especially convenient for many of the qtScoped* types.