Rename using, with
Rename our aforementioned pseudo-keywords to work around Visual Studio
being grumpy about the former. (The old names are more convenient, and
fairly safe since our "overload" is a function-macro and the native
keyword can't be used in that way, but it's dodgy enough that I'm not
quite willing to just tell VS to shut up.)

The new names are with_var and with_expr.