Fix build of Designer plugin on macOS
So... apparently, the "correct" way of building plugins for Designer
changed in Qt5, and, while an attempt was made to preserve compatibility
so that the Qt4 method would still work, it turns out it doesn't, for
some reason, on macOS.

Fortunately, it appears that porting to the new mechanism is possible.
Accordingly, adjust the include header to be the minimum needed header
(which is also the correct header for "the new way"), and adjust the
link libraries for the plugin so that a) including this header works for
both Qt4 and Qt5 and b) we avoid the broken shim header when building
against Qt5.

See also https://bugreports.qt.io/browse/QTBUG-73500.