Fix qtColorButton rendering with Oxygen
At some point, Oxygen decided to be clever and not draw LineEdit panels
if "the text won't fit". Alas, this happens to be the case for
qtColorButton, except of course we never actually draw text. Work around
Oxygen's "cleverness" by creating an artificial, very small sized font
whose metrics are passed in the style option structure in order to
"trick" Oxygen into believing the color swatch is not "too small".