Add helpers for creating meta-capable namespaces
Qt uses Q_ENUM_NS and Q_FLAG_NS to make meta-information available for
enumeration and flags types. The use of these macros requires use of the
Q_NAMESPACE macro. However, there are two serious flaws with this macro.
First, it can only appear once for a given namespace, requiring all
enumeration and flags types in a namespace to be declared in the same
header. Second, it is impossible to make the namespace usable outside of
the library which defines it. (This can be hacked around on ELF
platforms, but not on PE platforms, i.e. Windows.)

Qt 5.14, which is some time from release as of writing, will address the
second issue, but not the first. Since one of our users needs an
immediate fix, introduce helper macros to implement the work-around that
is used in Qt itself.