Improve AUTOMOC with qtExtensions
Since a432e9a14c8b, we provide helper macros that introduce a new
mechanism for adding meta-information. Unfortunately, CMake's AUTOMOC
does not recognize these out of the box, such that headers that use only
these helpers are not processed. Modify our configuration so that our
users will automatically have the key macro added to the set of macros
that CMake checks when deciding to run AUTOMOC on a header.