Add simple SVG widget
Create a simple widget to render an SVG. Unlike loading an SVG into a
QLabel, this supports resizing while still rendering at display
resolution (unlike QLabel, which renders to a pixmap once and then
resizes that), and unlike QSvgWidget, this supports changing the
position of the SVG within the widget.

Note that animation is not currently supported.