Fix immediate showing of combo box popup
qtComboBoxDelegate is intended to trigger the combo box popup
immediately (which saves users having to first trigger editing, then
separately trigger the popup, which is almost surely wanted immediately
since the user requested editing). This was not working, at least with
click-to-edit edit triggers; the popup would be shown, but immediately
hidden.

This appears to happen because the mouse-down event triggers editing,
which shows the popup. Then, at some later time, the user releases the
mouse button which causes QComboBox to hid the popup again. Because the
delay between events is from user input, simply queueing the popup won't
work.

The fix is subclass QComboBox in order to start a short timer when
showing the popup so that we can ignore hiding the popup if the mouse-up
event happens before the timer expires. This seems to effectively
prevent "regular clicks" from causing the problem, while a very long
click will still cause the popup to appear and disappear again. However,
this seems to be the appropriate behavior, since it allows the user to
press, move the mouse to select an item, and release to complete the
selection. (In fact, it seems that QComboBox is internally doing the
exact same thing, but because we are triggering the popup via code, we
miss QComboBox setting its own internal timer from a mouse-down event.)

As a stand-in for "a short time in which to differentiate between a
click and a long press", we are using the double-click interval (which,
again, is what QComboBox uses internally).