Add function to set icon theme
Create a new function in qtUtil to set the icon theme. See documentation
thereof for rationale. Additionally, due to an apparent Qt bug, getting
the name of the system theme requires a nasty work-around that involves
asking GTK(!), which makes it much more unreasonable for individual
applications to have to do this themselves.