Remove immutableness of type in test
It seems at some point, Qt became able to use types in containers that
cannot be assigned, and the natural sorting test was taking advantage of
this. However, the code does not compile with older versions of Qt as a
result. Make the type mutable, refactoring it to make the members
protected (and friending the debug helper that dumps the type, which is
the only thing that needs direct access). This makes the type
assignable, and is probably better anyway, as non-assignable types often
run into usage difficulties such as this.