Add QTE_THREAD_INVOKE
Add a helper macro to reexecute a method on the object's owning thread.
This is a paradigm that has been previously used in various places to
allow methods to be directly called from any thread when they need to
actually execute on the object's owning thread.

It used to be the case that doing this was moderately complicated,
involving lots of Q_ARG and other nastiness, but since Qt 5.10, this can
be done with lambdas in a way that is far more conducive to being
wrapped in a macro. Thus, it seems useful to make such a macro available
to our users.